+++ src/lib/api/stream.ts
@@ -9,7 +9,7 @@ const { App, Device, EventSource } = Plugins;
 import 'capacitor-eventsource';
 import { MessageResult, ErrorResult, EventSourcePlugin } from 'capacitor-eventsource';
 import { StreamData } from '../model/streamData';
-import ExtraInnings from '../extra-innings/extra-innings';
+import StreamEnhancer from './enhancers/stream-enhancer';
 import { Platform } from '@ionic/angular';
 
 const SECOND = 1000;
@@ -88,7 +88,7 @@ export class APIStream {
     console.debug(`APIStream(): default retry fallback: ${this.defaultRetryFallback}x`);
 
     const unenrichedSubject = new Subject<StreamData|ErrorEvent>();
-    this.subject = ExtraInnings.add(unenrichedSubject);
+    this.subject = StreamEnhancer.addEnhancements(unenrichedSubject);
 
     App.addListener('appStateChange', (state: AppState) => {
       if (this.deviceInfo?.platform !== 'web') {
